[fixup] refactor tests and address review comments
