+++ management-api-server/src/main/java/com/datastax/mgmtapi/resources/common/BaseResources.java
@@ -74,18 +74,22 @@ public abstract class BaseResources {
     }
   }
 
-  protected boolean keyspaceExists(String keyspaceName) {
+  /**
+   * Returns true if the specified keyspaceName is not null and a keyspace with the name exists.
+   * Returns false if the keyspaceName is null or if no keyspace with the name exists. Throws a
+   * ConnectionClosedException if there is an issue executing the RPC call to the Cassandra agent.
+   *
+   * @param keyspaceName The name of a keyspace you are looking for.
+   * @return True if the keyspace is found, false otherwise.
+   */
+  protected boolean keyspaceExists(String keyspaceName) throws ConnectionClosedException {
     if (keyspaceName != null) {
-      try {
-        ResultSet result =
-            app.cqlService.executePreparedStatement(
-                app.dbUnixSocketFile, "CALL NodeOps.getKeyspaces()");
-        Row row = result.one();
-        if (row != null) {
-          return row.getList(0, String.class).contains(keyspaceName);
-        }
-      } catch (Exception e) {
-        // cxonnection issue or some other problem. Eat it and return false.
+      ResultSet result =
+          app.cqlService.executePreparedStatement(
+              app.dbUnixSocketFile, "CALL NodeOps.getKeyspaces()");
+      Row row = result.one();
+      if (row != null) {
+        return row.getList(0, String.class).contains(keyspaceName);
       }
     }
     return false;
